% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_date_parse.R
\name{safe_date_parse}
\alias{safe_date_parse}
\title{Safely Parse Dates}
\usage{
safe_date_parse(x)
}
\arguments{
\item{x}{A character vector containing dates.}
}
\value{
A Date vector.
}
\description{
Converts mixed-format date vectors into standard Date objects.
This function helps simplify parsing inconsistent date formats
commonly found in messy datasets.
}
\examples{
safe_date_parse(c("2024-01-01", "01/02/2024"))

}
